Rename: post 공통 컴포넌트 구조 수정
