+++ src/app/(main)/activity/[semesterId]/[activityId]/create-board/_components/detail/BoardDetail.tsx
