+++ src/app/(main)/activity/[semesterId]/[activityId]/create-board/_components/detail/index.ts
@@ -1,0 +1,1 @@
+export { CreateBoardDetail } from './BoardDetail'
