+++ src/app/(main)/activity/[semesterId]/[activityId]/create-board/_components/form/form-field/index.ts
@@ -1,0 +1,1 @@
+export { BoardFormField } from './FormField'
