+++ src/app/(main)/activity/[semesterId]/[activityId]/create-board/_components/form/select-member-input/SelectMemberInput.tsx
@@ -8,7 +8,7 @@ import { Skeleton } from '@/components/ui'
 import { UserQuries } from '@/service/api'
 import { UserResponseDto } from '@/service/models'
 
-import { MultipleMemberSelect } from './MultipleMemberSelect'
+import { MultipleMemberSelect } from './multiple-member-select'
 
 type SelectMemberInputProps = {
   selectedMember: UserResponseDto[]
