+++ src/app/(main)/activity/[semesterId]/[activityId]/create-board/_components/form/select-member-input/index.ts
@@ -1,0 +1,1 @@
+export { SelectMemberInput } from './SelectMemberInput'
