+++ src/app/(main)/activity/[semesterId]/[activityId]/create-board/_components/form/select-member-input/multiple-member-select/index.ts
@@ -1,0 +1,1 @@
+export { MultipleMemberSelect } from './MultipleMemberSelect'
