+++ src/app/(main)/activity/[semesterId]/[activityId]/create-board/_components/hero/BoardHero.tsx
