+++ src/app/(main)/activity/[semesterId]/[activityId]/create-board/_components/hero/index.ts
@@ -1,0 +1,1 @@
+export { CreateBoardHero } from './BoardHero'
