+++ src/app/(main)/event/board/create-post/_components/form/Form.tsx
@@ -4,7 +4,7 @@ import { useForm } from 'react-hook-form'
 
 import { zodResolver } from '@hookform/resolvers/zod'
 
-import { CreatePostForm } from '@/components/feature'
+import { Form } from '@/components/ui'
 import { CreateEventPost, CreateEventPostSchema } from '@/schema/post'
 
 export const CreateEventPostForm = () => {
@@ -17,11 +17,10 @@ export const CreateEventPostForm = () => {
   })
 
   return (
-    <CreatePostForm
-      form={form}
-      onSubmit={() => {}}
-      isExecuting={false}
-      isImageRequired={false}
-    />
+    <Form {...form}>
+      <form>
+        <div>temp form ui</div>
+      </form>
+    </Form>
   )
 }
