--- src/components/feature/post/create-post-form/index.ts
@@ -1,1 +1,0 @@
-export { CreatePostForm } from './CreatePostForm'
