--- src/components/feature/post/form-field/index.ts
@@ -1,1 +1,0 @@
-export { PostFormField } from './FormField'
