+++ src/components/feature/post/index.ts
@@ -1,5 +1,4 @@
-export * from './create-post-form'
-export * from './form-field'
+export * from './post-editor'
 export * from './table'
 export * from './content'
 export * from './image-input'
